# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"LoginResponse": {
				"properties": {
					"token": {
						"type": "string"
					},
					"requiereCompletarPerfil": {
						"type": "boolean"
					},
					"user": {
						"properties": {
							"empresaId": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"perfilCompleto": {
								"type": "boolean"
							},
							"rolEquipo": {
								"type": "string"
							},
							"tipoUsuario": {
								"type": "string",
								"nullable": true
							},
							"rol": {
								"type": "string"
							},
							"correo": {
								"type": "string"
							},
							"username": {
								"type": "string"
							},
							"idUsuario": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"empresaId",
							"perfilCompleto",
							"rolEquipo",
							"tipoUsuario",
							"rol",
							"correo",
							"username",
							"idUsuario"
						],
						"type": "object"
					}
				},
				"required": [
					"token",
					"requiereCompletarPerfil",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"correo": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"correo",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SolicitudResetDTO": {
				"properties": {
					"correo": {
						"type": "string"
					}
				},
				"required": [
					"correo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConfirmacionResetDTO": {
				"properties": {
					"token": {
						"type": "string"
					},
					"nuevaContrasena": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"nuevaContrasena"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"legacyJwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "producttrack",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@mariaximena0724 y emilia"
		}
	},
	"paths": {
...
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Token de autenticación y datos del usuario.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					}
				},
				"description": "Inicia sesión en el sistema con correo y contraseña.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/log.controller)",
				"tags": [
					"Autenticación"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Datos de inicio de sesión.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest",
								"description": "Datos de inicio de sesión."
							}
						}
					}
				}
			}
		},
		"/auth/solicitar-reset": {
			"post": {
				"operationId": "SolicitarReset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensaje": {
											"type": "string"
										}
									},
									"required": [
										"mensaje"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Solicita el restablecimiento de contraseña.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/log.controller)",
				"tags": [
					"Autenticación"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SolicitudResetDTO"
							}
						}
					}
				}
			}
		},
		"/auth/confirmar-reset": {
			"post": {
				"operationId": "ConfirmarReset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensaje": {
											"type": "string"
										}
									},
									"required": [
										"mensaje"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Confirma el restablecimiento de contraseña con el token recibido.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/log.controller)",
				"tags": [
					"Autenticación"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConfirmacionResetDTO"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}